% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnet.R
\name{cnet}
\alias{cnet}
\title{Plots the canonical netowrk map of classified regions}
\usage{
cnet(x, n, thres, ...)
}
\arguments{
\item{x}{regs object}

\item{n}{number of regions}

\item{thres}{the cross-corelation threshold of the network}
}
\value{
A map with the canonical network
}
\description{
Plots the canonical netowrk map of classified regions
}
\details{
ToDo
}
