% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_to_som.R
\name{dt_to_som}
\alias{dt_to_som}
\title{Transform a data table object in tidy format to a matrix object}
\usage{
dt_to_som(dataset)
}
\arguments{
\item{dataset}{A four-column data table object: time, latitude, longitude and variable}

\item{dir}{Name of the directory that the new objects will be saved}
}
\value{
A list with a matrix that can be used as input for the som() function of
the kohonen package (input_for_som), a data table linking ids with latitude and
longitude (coords) and the original dataset (input_dt)
}
\description{
Transform a data table object in tidy format to a matrix object
}
\examples{
dt_to_som(my_tidy_dt)
}
