% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owda.R
\docType{data}
\name{owda}
\alias{owda}
\title{Old World Drought Atlas (1500-2012)}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 1355264 rows and 4 columns.
}
\source{
\href{https://www.fzp.czu.cz/en/r-9409-science-research/r-9674-leading-research-groups/r-9669-hydrological-and-climate-variability/r-9713-team-news/webapp-for-hydroclimate-reconstruction.html}{FZP archive}
}
\usage{
data(owda)
}
\description{
Reconstruction of European hydroclimate derived from tree-rings.
The variable used is self-calibrated Palmer Drought Severity Index (scPDSI)
at annual time step.
}
\examples{
str(owda)
}
\references{
Markonis et al. (2018) Nature Communications 9(1):1767
(\href{https://www.nature.com/articles/s41467-018-04207-7}{Nature Springer})
}
\keyword{datasets}
