% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clusters.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plots classification maps}
\usage{
plot_clusters(som_clusters, clusters, nrow, ncol, ...)
}
\arguments{
\item{som_clusters}{Data table created by som_clusters()}

\item{clusters}{Subset of clusters to be ploted}

\item{nrow}{Number of plots in each column}

\item{ncol}{Number of plots in each row}
}
\value{
Plots the classfication clusters, representing homogeneous regions of SOM in space
}
\description{
Plots classification maps
}
\examples{
plot_clusters(my_som_clusters)

}
