% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_somspace.R
\name{plot_som_space}
\alias{plot_som_space}
\title{Plots SOM map}
\usage{
plot_som_space(som_data, ...)
}
\arguments{
\item{som_data}{List created by som_map}
}
\value{
Plots the nodes of SOM in space
}
\description{
Plots SOM map
}
\examples{
plot_som_map(my_som)

}
