% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_som_summary.R
\name{plot_som_summary}
\alias{plot_som_summary}
\title{Plots SOM maps with node characteristics}
\usage{
plot_som_summary(som_data, ...)
}
\arguments{
\item{som_data}{List returned by dt_to_som}
}
\description{
Plots SOM maps with node characteristics
}
\examples{
plot_som_summary(my_som)

}
