% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regs_class.R
\docType{data}
\name{regs}
\alias{regs}
\title{Spatial SOM regions class}
\format{An object of class \code{regs} of length 0.}
\usage{
regs
}
\description{
Spatial SOM regions class
}
\details{
ToDo
}
\examples{
my_som <- somspa(aa), relen = 1000, grid = somgrid(6, 6, "hexagonal"))
summary(my_som)
plot(my_som)

}
\keyword{datasets}
