% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somsp_class.R
\docType{data}
\name{somsp}
\alias{somsp}
\title{Spatial SOM object class}
\format{An object of class \code{somsp} of length 0.}
\usage{
somsp
}
\description{
Spatial SOM object class
}
\details{
ToDo
}
\examples{
my_som <- somspa(aa), relen = 1000, grid = somgrid(6, 6, "hexagonal"))
summary(my_som)
plot(my_som)

}
\keyword{datasets}
