% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ts.R
\name{plot_ts}
\alias{plot_ts}
\title{Plot time series}
\usage{
plot_ts(x, n)
}
\arguments{
\item{x}{is either a \code{somsp} or a \code{regs} object}

\item{n}{is either the set of nodes for \code{somsp} or the number of regions for \code{regs}}
}
\value{
plot object
}
\description{
Plots the time series of SOM nodes or regions mean
}
\details{
In case of \code{regs}, all the regions are ploted.
}
\seealso{
\code{\link{somspa}}
}
