% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regs_class.R
\docType{data}
\name{regs}
\alias{regs}
\title{Regions class}
\format{
An object of class \code{regs} of length 0.
}
\usage{
regs
}
\description{
Regions class
}
\details{
The \code{regs} class contains:

\itemize{

\item{A summary \code{data.table} which updates the \code{somsp} object with the region ids of all classification schemes
up to \code{nregions}. Each different classification scheme is stored as an individual region, e.g. \code{regions.2},
\code{regions.3}, etc.}

\item{A \code{data.table} with the original data set, as in \code{somsp}.}
}

It can be plotted by \code{plot} and \code{plot_ts}.
If \code{plot} is used, three additional arguments are needed; a set with the classification schemes
that will be ploted, number of rows and number of columns of the plotted panels.
\code{plot_ts} plots all the time series of a given classification scheme.
}
\seealso{
\code{\link{somsp}}

\code{\link{somregs}}
}
\keyword{datasets}
