% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sominp.R
\name{sominp}
\alias{sominp}
\title{Create sominp object}
\usage{
sominp(x, ...)
}
\arguments{
\item{x}{The \code{data.table} object which will be tranformed to \code{somsp} object.}
}
\description{
\code{sominp} transforms the data set from \code{data.table} to
\code{somsp} format, which can be used as argument in the \code{somspa} function.
}
\details{
\code{x} should be in \href{https://cran.r-project.org/web/packages/tidyr/vignettes/tidy-data.html}{tidy format}
and contain four columns: time, latitude, longitude and variable.
}
\examples{
\dontrun{}
som_input(my_tidy_dt)

}
