% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sominp.R
\name{sominp}
\alias{sominp}
\title{Transform a data table object in tidy format to a somsp object}
\usage{
sominp(x, ...)
}
\arguments{
\item{x}{The data table object that will be tranformed}
}
\value{
A sominp object. It contains a matrix that can be used as input for the som() function of
the kohonen package (input_for_som), a data table linking ids with latitude and
longitude (coords) and the original dataset (input_dt)
}
\description{
Transform a data table object in tidy format to a somsp object
}
\details{
x should contain four columns: time, latitude, longitude and variable
}
\examples{
som_input(my_tidy_dt)

}
