% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somsp_class.R
\docType{data}
\name{somsp}
\alias{somsp}
\title{Spatial SOM class}
\format{
An object of class \code{somsp} of length 0.
}
\usage{
somsp
}
\description{
Spatial SOM class
}
\details{
The \code{somsp} objects are created by \code{somspa} function and contain:

\itemize{

\item{A summary \code{data.table} with the coordinates of each SOM node, the distances of objects
to their corresponding winning unit, the number of points of each node, as well as the median
latitude and longitude of each node coordinates and their standard deviation.}

\item{A Self-Organizing Map object (see also \code{\link{kohonen}}).}

\item{The \code{sominp} object used as input for the SOM, with an id number coressponding to
location and a node number to the classification group of SOM.}
}

They can be plotted by \code{plot} and \code{plot_ts} functions or summarized by \code{summary}.
}
\seealso{
\code{\link{somspa}}

\code{\link{plot_ts}}
}
\keyword{datasets}
