% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somspa.R
\name{somspa}
\alias{somspa}
\title{Transform a data table object in tidy format to a matrix object}
\usage{
somspa(x, ...)
}
\arguments{
\item{x}{A somsp object}

\item{...}{Arguments of som() of kohonen package}
}
\value{
A list of a data table and a som object

\itemize{

\item{summary}{coordinates of each SOM node, the distances of objects
to their corresponding winning unit, the number of points of each node, as well as the median
latitude and longitude of each node coordinates, as well as their standard deviation}

\item{som}{Self-Organizing Map (see \code{\link{som}})}
}
}
\description{
Transform a data table object in tidy format to a matrix object
}
\details{
ToDo
}
\examples{
input_for_som <- sominp(my_dt)
my_som <- somspa(aa), relen = 1000, grid = somgrid(6, 6, "hexagonal"))
my_som$summary
my_som$som

}
\seealso{
\code{\link{kohonen}}
}
